requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
werkzeug==2.3.7
flask-jwt-extended==4.5.3
requests==2.31.0
pytest==7.4.0
python-dotenv==1.0.0
clarifai-grpc==9.10.4
pillow==10.0.0